[refactor(socket)] 將socket相關邏輯獨立成一個檔案以避免循環import
